refactor: 강의 모듈 불필요한 의존성 분리, DTO 컨벤션 적용, 강의 service 불필요한 코드 삭제
