+++ usw-suwiki-core/usw-suwiki-version/src/main/java/usw/suwiki/core/version/v1/VersionController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import usw.suwiki.common.response.ResponseForm;
-import usw.suwiki.domain.lecture.service.LectureCRUDService;
+import usw.suwiki.domain.lecture.service.LectureService;
 
 import java.util.List;
 
@@ -18,21 +18,21 @@ import java.util.List;
 @RestController
 @RequiredArgsConstructor
 public class VersionController {
-    private final LectureCRUDService lectureCRUDService;
+  private final LectureService lectureService;
 
-    @GetMapping("/version")
-    public ResponseEntity<VersionResponseDto> findVersionSuwiki() {
-        HttpHeaders header = new HttpHeaders();
-        float version = SuwikiVersion.version;
-        VersionResponseDto dto = new VersionResponseDto(version);
-        return new ResponseEntity<>(dto, header, HttpStatus.valueOf(200));
-    }
+  @GetMapping("/version")
+  public ResponseEntity<VersionResponseDto> findVersionSuwiki() {
+    HttpHeaders header = new HttpHeaders();
+    float version = SuwikiVersion.version;
+    VersionResponseDto dto = new VersionResponseDto(version);
+    return new ResponseEntity<>(dto, header, HttpStatus.valueOf(200));
+  }
 
-    @GetMapping("/majorType")
-    public ResponseEntity<ResponseForm> findAllMajorType() {
-        HttpHeaders header = new HttpHeaders();
-        List<String> list = lectureCRUDService.loadMajorTypes();
-        ResponseForm data = new ResponseForm(list);
-        return new ResponseEntity<>(data, header, HttpStatus.valueOf(200));
-    }
+  @GetMapping("/majorType")
+  public ResponseEntity<ResponseForm> findAllMajorType() {
+    HttpHeaders header = new HttpHeaders();
+    List<String> list = lectureService.loadMajorTypes();
+    ResponseForm data = new ResponseForm(list);
+    return new ResponseEntity<>(data, header, HttpStatus.valueOf(200));
+  }
 }
