+++ usw-suwiki-lecture/build.gradle
@@ -9,14 +9,13 @@ allprojects {
 project(':usw-suwiki-lecture') {
     dependencies {
         implementation project(':usw-suwiki-user')
-
-        implementation 'com.googlecode.json-simple:json-simple:1.1.1'
     }
 }
 
 project(':usw-suwiki-lecture:usw-suwiki-schedule') {
     dependencies {
         implementation project(':usw-suwiki-lecture')
+        implementation 'com.googlecode.json-simple:json-simple:1.1.1'
     }
 }
