--- usw-suwiki-lecture/src/main/java/usw/suwiki/domain/lecture/dto/LectureAndCountResponseForm.java
@@ -1,11 +1,0 @@
-package usw.suwiki.domain.lecture.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class LectureAndCountResponseForm {
-    Object data;
-    Long count;
-}
