--- usw-suwiki-lecture/src/main/java/usw/suwiki/domain/lecture/dto/LectureFindOption.java
@@ -1,16 +1,0 @@
-package usw.suwiki.domain.lecture.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class LectureFindOption {
-    private String orderOption;
-    private Integer pageNumber;
-    private String majorType;
-
-    public boolean passMajorFiltering() {
-        return majorType == null || majorType.equals("전체") || majorType.isEmpty();
-    }
-}
