+++ usw-suwiki-lecture/src/main/java/usw/suwiki/domain/lecture/dto/Lectures.java
@@ -1,0 +1,11 @@
+package usw.suwiki.domain.lecture.dto;
+
+import usw.suwiki.domain.lecture.Lecture;
+
+import java.util.List;
+
+public record Lectures(
+  List<Lecture> content,
+  Long count
+) {
+}
