+++ usw-suwiki-lecture/src/main/java/usw/suwiki/domain/lecture/dto/TimetableResponse.java
@@ -1,34 +1,40 @@
 package usw.suwiki.domain.lecture.dto;
 
 import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import usw.suwiki.domain.lecture.timetable.Timetable;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.util.List;
 
-@Getter
-@Builder
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class TimetableResponse {
 
-    long id;
-    int year;
-    String semester;
-    String name;
-    List<TimetableCellResponse> cellList;
+  @Data
+  public static class Detail {
+    private final Long cellId;
+    private final String name;
+    private final int year;
+    private final String semester;
+    private final List<Cell> cells;
+  }
 
-    public static TimetableResponse of(Timetable timetable) {
-        return builder()
-            .id(timetable.getId())
-            .year(timetable.getYear())
-            .semester(timetable.getSemester().getValue())
-            .name(timetable.getName())
-            .cellList(
-                timetable.getCellList().stream()
-                    .map(TimetableCellResponse::of)
-                    .toList())
-            .build();
-    }
+  @Data
+  public static class Simple {
+    private final Long cellId;
+    private final String name;
+    private final int year;
+    private final String semester;
+  }
+
+  @Data
+  public static class Cell {
+    private final Long cellId;
+    private final String lecture;
+    private final String professor;
+    private final String color;
+    private final String location;
+    private final String day;
+    private final int startPeriod;
+    private final int endPeriod;
+  }
 }
